Add files via upload
Added Module 14 Day 1 Project:
Added a .CSV text file that contains a person's name, grade, and semester which is used in the Mod 14 Day 1 project.
The project reads a CSV file called sample_grades.csv and determines the mean, median, and standard deviation from a list of grades for each semester. It outputs the results to the screen.